Add a --start_index option
This adds an offset to the index in filenames generated by the design step. The intention is that we can run multiple parallel design batches, pass these to inverse_folding, folding and design_folding as distinct batches, then merge into a single directory tree without filename clobbering to run analysis and filtering steps.